поправил логи, добавил поле в модель заказа на продажу

+++ Backend/SteamTradeBot.Backend/Services/OrderCancellingService.cs
@@ -19,13 +19,23 @@ public class OrderCancellingService
     public async Task ClearBuyOrdersAsync(string apiKey)
     {
         Log.Logger.Information("Start buy orders canceling...");
-        var itemsWithPurchaseOrders = await _marketDb.GetBuyOrdersAsync(apiKey);
-        foreach (var order in itemsWithPurchaseOrders)
+        var buyOrders = await _marketDb.GetBuyOrdersAsync(apiKey);
+        foreach (var order in buyOrders)
         {
-            Log.Logger.Information("Cancelling item {0} with buy price {1}", order.EngItemName, order.BuyPrice);
-            await _api.CancelBuyOrderAsync(order.ItemUrl);
-            await _marketDb.RemoveBuyOrderAsync(order);
-            Log.Logger.Information("Item {0} with buy price {1} was canceled", order.EngItemName, order.BuyPrice);
+            Log.Logger.Information("Cancelling item {0} (Price: {1}, Quantity: {2})", 
+                order.EngItemName, order.BuyPrice, order.Quantity);
+            var successfullyCanceled = await _api.CancelBuyOrderAsync(order.ItemUrl);
+            if (successfullyCanceled)
+            {
+                await _marketDb.RemoveBuyOrderAsync(order);
+                Log.Logger.Information("Item {0} (Price: {1}, Quantity: {2}) was canceled", 
+                    order.EngItemName, order.BuyPrice, order.Quantity);
+            }
+            else
+            {
+                Log.Logger.Warning("Can't cancel item {0} (Price: {1}, Quantity: {2})", 
+                    order.EngItemName, order.BuyPrice, order.Quantity);
+            }
         }
         Log.Logger.Information("All buy orders have been canceled!");
     }
